# apps/next-app-zod/public/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server",
      "name": "local"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "ApiResponseSchema": {
        "type": "object",
        "discriminator": {
          "propertyName": "status"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/SuccessResponseSchema"
          },
          {
            "$ref": "#/components/schemas/ErrorResponseSchema"
          }
        ]
      },
...
      "CreateSubscriptionBody": {
        "type": "object",
        "properties": {
          "callbackUrl": {
            "type": "string",
            "format": "uri",
            "description": "HTTPS URL that will receive events"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "payment.succeeded",
                "payment.failed",
                "payment.refunded"
              ]
            },
            "minItems": 1,
            "description": "Event types this subscription receives"
          },
          "secret": {
            "type": "string",
            "minLength": 32,
            "description": "Shared HMAC secret used to sign deliveries"
          }
        },
        "required": [
          "callbackUrl",
          "events",
          "secret"
        ],
        "additionalProperties": false
      },
...
      "EmailNotificationSchema": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "email"
            ]
          },
          "to": {
            "type": "string",
            "format": "email",
            "description": "Recipient email address"
          },
          "subject": {
            "type": "string",
            "minLength": 1,
            "description": "Email subject"
          },
          "body": {
            "type": "string",
            "description": "Email body content"
          }
        },
        "required": [
          "type",
          "to",
          "subject",
          "body"
        ]
      },
      "ErrorResponseSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "error"
            ]
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "number",
            "description": "Error code"
          }
        },
        "required": [
          "status",
          "error",
          "code"
        ]
      },
...
      "NotificationSchema": {
        "type": "object",
        "discriminator": {
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/EmailNotificationSchema"
          },
          {
            "$ref": "#/components/schemas/SmsNotificationSchema"
          },
          {
            "$ref": "#/components/schemas/PushNotificationSchema"
          }
        ]
      },
...
      "PushNotificationSchema": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "push"
            ]
          },
          "deviceId": {
            "type": "string",
            "description": "Device identifier for push notification"
          },
          "title": {
            "type": "string",
            "description": "Notification title"
          },
          "body": {
            "type": "string",
            "description": "Notification body"
          }
        },
        "required": [
          "type",
          "deviceId",
          "title",
          "body"
        ]
      },
...
      "SmsNotificationSchema": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "sms"
            ]
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number with country code (e.g., +1234567890)"
          },
          "message": {
            "type": "string",
            "maxLength": 160,
            "description": "SMS message (max 160 characters)"
          }
        },
        "required": [
          "type",
          "phoneNumber",
          "message"
        ]
      },
...
      "SubscriptionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "ulid",
            "description": "Subscription identifier"
          },
          "callbackUrl": {
            "type": "string",
            "format": "uri",
            "description": "Registered callback URL"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "description": "Subscription creation time"
          }
        },
        "required": [
          "id",
          "callbackUrl",
          "createdAt"
        ]
      },
      "SuccessResponseSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Unique identifier"
              },
              "message": {
                "type": "string",
                "description": "Success message"
              }
            },
            "required": [
              "id",
              "message"
            ]
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "examples": [
                    false
                  ]
                },
                "error": {
                  "type": "string",
                  "examples": [
                    "Validation error"
                  ]
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "examples": [
                    false
                  ]
                },
                "error": {
                  "type": "string",
                  "examples": [
                    "Something went wrong"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/integrations/subscribe": {
      "post": {
        "operationId": "zodCreateSubscription",
        "summary": "Create subscription",
        "description": "Registers a callback URL that will receive signed payment events.",
        "tags": [
          "Integrations",
          "Webhooks"
        ],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "servers": [
          {
            "url": "https://api.example.com/v1,",
            "description": "https://api-eu.example.com/v1"
          }
        ],
        "externalDocs": {
          "url": "https://docs.example.com/integrations/webhooks",
          "description": "Webhook signing guide"
        },
        "callbacks": {
          "paymentEvent": {
            "{$request.body#/callbackUrl}": {
              "$ref": "#/components/callbacks/SubscriptionEventPayload"
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionBody"
              }
            }
          },
          "description": "Subscription registration payload"
        },
        "responses": {
          "201": {
            "description": "Subscription created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL of the created subscription",
                "schema": {
                  "type": "string"
                }
              }
            },
            "links": {
              "getSubscription": {
                "operationId": "zodGetSubscription"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "operationId": "get-notifications",
        "summary": "Get notification status",
        "description": "Returns success or error response (demonstrates Zod discriminated union)",
        "tags": [
          "Notifications"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-notifications",
        "summary": "Send a notification",
        "description": "Send notification via email, SMS, or push (demonstrates Zod discriminated unions with validation)",
        "tags": [
          "Notifications"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Invalid notification data or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSchema"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}